+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Minor: Added setting for character limit of deleted messages. (#6491)
 - Minor: Added link support to plugin message API. (#6386)
 - Minor: Added a description for the logging option under moderation tab. (#6514)
+- Minor: Added a menu action to sort tabs alphabetically.
 - Bugfix: Expose the "Extra extension IDs" setting on non-Windows systems too. (#6509)
 - Bugfix: Fixed scrollbar rect computation potentially resulting in overflows. (#6547)
 - Dev: Update release documentation. (#6498)